feat: _check_validity_on_delete support
In json files _check_validity on support can be added at the start of the file (before features, next to dataset). If the value is false the current validity is not validated, and not required.

Furthermore changed dataset-name to dataset for easy metadata merging, and fixed a empty id collection bug in persistence cache loading