fix: invalidate flattened nested features
Because flattening happends before validation, nested featured which we created ourselve appeared in the timeline. This is solved by keeping a set of invalid feature_stream, which we check during validation.
This is maybe not the best solution, however flatten first first and validate after is easier for double nested features. At least in my head right now ;)