refac: no join on replay
We uses a join on last-n-features for parent-collection and parent-id. This seems to be really slow for big datasets (bgt plaatsbepalingspunten for example). We already saved the previous_version, now without the join we can just doe a sequential scan. Per batch we enrich the stream data with parent-collection and parent-id. This does add some query overhead, but performance overall should be better (without DB optimizations)