Use geometry attribute for all (#17)
* feat: support extra geomerty-attributes

* support function json-reader: ~#geo-attr [PDOK-3512]

* geo-attribute has to be delivered as a map (internal geo representation)

* Format test resource

To analyse test errors test resources are formatted to see what will
be tested

[PDOK-3512]

* Use Java class GeometryAttribute instead of Clojure geometry-map

Beside the main-geometry the json-input can contain an extra
geometry-attribute. To serialize this attribute, make these attributes
persistence in database and have a consistent codebase the geometries
will be coded in a Java-class GeometeryAttribute.

[PDOK-3512]

* Removed use of meta(info) :geo-attr

The feature to use extra geometry-attribute was implemented
with the use of the meta(info)-tag :geo-attr. Now will this
feauture be implemented with the use of GeometryAttribute class.

[PDOK-3512]

* Extend function as-ga, so tiles will be set add GeometryAttribute

[PDOK-3512]

* Use of renamed function as-geometry-attribute --> as-ga

[PDOK-3512]

* Made input-format of extra geometry(attribute) equal to main geometry

To make input-format of all input geometry consisent, so we use the
format of the alreay existing main-geometry. This is a map with the outline
("type":type, type geometry).

[PDOK-3512]

* Renamed the function #geo-attr to #geometry

[PDOK-3512]

* Use GeometryAttribute(.class) as data-type for main-geometry

To make code more consistent al geometries have to be transformed
to a GeometryAttribute. For this type is a serialization available
in featured-shared.

[PDOK-3512]

* Use in test GeometryAttribute instead of geometry-map

[PDOK-3512]

* Removed function as-ga from namespace util

The function util/as-ga is not more necessary because
GeometryAttribute is everywhere used as the datatype for
geometry.

[PDOK-3512]

* Use released version of featured-shared with GeometryAttribute

[PDOK-3512]

* Cleanup: removed some test files

[PDOK-3512]

* Cleanup: removed some test files

[PDOK-3512]

* Moved clj-to-pg-type to the namespace postgres

[PDOK-3512, review]

* Add test file with some features

[PDOK-3512]

* Use update-in to transform main-geometry in attribute-list

[PDOK-3512, review comment]

* Use update-in to transform main-geometry in attribute-list

[PDOK-3512, review comment]

* Use postwalk to create GeometryAttribute objects

[PDOK-3512]

* One function to create columns (refactoring)

[PDOK-3512 review]

* Better naming voor function to create GeometryAttribute

[PDOK-3512 review]

* Better naming function replace-geometry --> update-geometry

[PDOK-3512 review]

* Some logging removed

This statement will log too many entries, so removed this line.

[PDOK-3512 review]

* Use the function transit/to-json inline

To make code more simple the function transit/to-json is used
with with update function, instead of the assoc function.

[PDOK-3512 review comments]

* Splitted the create-column function in execute-sql and create the sql-string

[PDOK-3512 review]