New or change (#18)
For BAG we don't know if it is a new or a change. To solve this we introduce a new action "new|change", which is resolved to either a new or a change in the processor. The feature stream contains the result (new or change) and not the original new|change.